+++ lib/food/index.js
@@ -2,6 +2,7 @@
 
 const request = require('request');
 const haksik = require('./haksik');
+const parser = require('./parsers');
 const {Iconv} = require('iconv');
 const errors = require('./errors');
 
@@ -12,7 +13,7 @@ async function getOfficer(dayOfWeek) {
   return new Promise((resolve, reject) => {
     request(haksik.getSoonguriJsonUrl(dayOfWeek), (err, response, body) => {
       if (!err) {
-        resolve(haksik.parseOfficer(body));
+        resolve(parser.parseOfficer(body));
       } else {
         reject(new errors.SoongguriJsonConnectionError());
       }
@@ -25,7 +26,7 @@ async function getPupil(dayOfWeek) {
   return new Promise((resolve, reject) => {
     request(haksik.getSoonguriJsonUrl(dayOfWeek), (err, response, body) => {
       if (!err) {
-        resolve(haksik.parsePupil(body));
+        resolve(parser.parsePupil(body));
       } else {
         reject(new errors.SoongguriJsonConnectionError());
       }
@@ -52,7 +53,7 @@ async function getDorm(option) {
     }, (err, response) => {
       if (!err) {
         const resEucKr = iConv.convert(response.body).toString();
-        resolve(haksik.parseDorm(resEucKr));
+        resolve(parser.parseDorm(resEucKr));
       } else {
         reject(new errors.SsuDormConnectionError());
       }
